+++ Tests/SentryTests/Integrations/Performance/AppStartTracking/SentryAppStartTrackingIntegrationTests.swift
@@ -51,6 +51,19 @@ class SentryAppStartTrackingIntegrationTests: XCTestCase {
         XCTAssertNil(SentrySDK.getAppStartMeasurement())
     }
     
+    func testHybridSDKModeEnabled_DoesUpdatesAppState() {
+        PrivateSentrySDKOnly.appStartMeasurementHybridSDKMode = true
+        
+        let options = fixture.options
+        options.tracesSampleRate = 0.0
+        options.tracesSampler = nil
+        sut.install(with: options)
+        
+        TestNotificationCenter.uiWindowDidBecomeVisible()
+        
+        XCTAssertNotNil(SentrySDK.getAppStartMeasurement())
+    }
+    
     func testOnlyAppStartMeasuringEnabled_DoesNotUpdatesAppState() {
         let options = fixture.options
         options.tracesSampleRate = 0.0